+++ db/migrations/20191006162403_create-process-table.js
@@ -1,4 +1,4 @@
-exports.up = function(knex, Promise) {
+exports.up = function(knex) {
   return knex.schema.createTable("process", table => {
     table.uuid("id").primary();
     table.uuid("workflow_id").notNullable();
@@ -8,6 +8,6 @@ exports.up = function(knex, Promise) {
   });
 };
 
-exports.down = function(knex, Promise) {
+exports.down = function(knex) {
   return knex.schema.dropTable("process");
 };