+++ db/migrations/20210510165700_add_blueprint_hash_to_workflow.js
@@ -1,12 +1,11 @@
-
 exports.up = function (knex) {
-    return knex.schema.table('workflow', table => {
-        table.text('blueprint_hash')
-    });
+  return knex.schema.table("workflow", (table) => {
+    table.text("blueprint_hash");
+  });
 };
 
 exports.down = function (knex) {
-    return knex.schema.table('workflow', table => {
-        table.dropColumn('blueprint_hash');
-    });
+  return knex.schema.table("workflow", (table) => {
+    table.dropColumn("blueprint_hash");
+  });
 };