+++ db/seeds/packages/test_package.js
@@ -1,18 +1,15 @@
-const test_workflow_package =
-  ["do",
-
-  [ "js", ["`", "function lisp_test_task(args) { \
-    const n_interp = args[0].n_interp;\
-    let dates = [];\
-    for(let i = 0; i<n_interp; i++){\
-      dates.push(new Date());\
-    };\
-    const result = {dates: dates};\
-    return result;  \
-  }"]],
-
-  null];
+const test_workflow_package = [
+  "do",
+  [
+    "js",
+    [
+      "`",
+      "function lisp_test_task(args) { const n_interp = args[0].n_interp; let dates = []; for(let i = 0; i<n_interp; i++){ dates.push(new Date()); }; const result = {dates: dates}; return result; }",
+    ],
+  ],
+  null,
+];
 
 module.exports = {
-  test_workflow_package
-}
+  test_workflow_package,
+};