+++ src/middlewares/actordata.js
@@ -36,14 +36,15 @@ const captureActorData = async (ctx, next) => {
   const session_id = ctx.state.user.session_id;
 
   if (ctx.state.actor_data) {
-    (ctx.state.actor_data["actor_id"] = actor_id), (ctx.state.actor_data["claims"] = claims);
+    ctx.state.actor_data["actor_id"] = actor_id;
+    ctx.state.actor_data["claims"] = claims;
     ctx.state.actor_data.trace = trace;
     ctx.state.actor_data["session_id"] = session_id;
   } else {
     ctx.state.actor_data = { trace, actor_id, claims, session_id };
   }
 
-  return await next();
+  return next();
 };
 
 module.exports = {