+++ src/utils/engineListener.js
@@ -3,11 +3,28 @@ const { logger } = require("./logger");
 const mqtt = require("../services/mqtt");
 
 const processStateListener = async (processState) => {
-  logger.info(`PS LISTENER: PID [${processState.id}]`);
+  if(process.env.PUBLISH_STATE_EVENTS) {
+    const topic = `/process/${processState.process_id}/state`;
+
+    const message = {
+      stateId: processState.id,
+      processId: processState.process_id,
+      stepNumber: processState.step_number,
+      nodeId: processState.node_id,
+      status: processState.status,
+      workflow: processState.workflow_name
+    };
+  
+    mqtt.publishMessage(topic, message);
+    logger.info(`PS LISTENER: PID [${processState.id}], step [${processState.step_number}], status [${processState.status}]`);
+  }
 };
 
 const activityManagerListener = async (activityManager) => {
   logger.info(`AM LISTENER: AMID [${activityManager._id}]`);
+  if (!activityManager._id) {
+    return
+  }
 
   const topic = `/process/${activityManager._process_id}/am/create`;
 
@@ -20,14 +37,14 @@ const activityManagerListener = async (activityManager) => {
 
   mqtt.publishMessage(topic, message);
 
-  if (activityManager._props.result.session_id) {
+  if (activityManager?._props?.result?.session_id) {
     const sessionTopic = `/session/${activityManager._props.result.session_id}/am/create`;
     mqtt.publishMessage(sessionTopic, message);
   } else {
     logger.info("AM LISTENER: No session provided");
   }
 
-  if (activityManager._props.result.actor_id) {
+  if (activityManager?._props?.result?.actor_id) {
     const actorTopic = `/actor/${activityManager._props.result.actor_id}/am/create`;
     mqtt.publishMessage(actorTopic, message);
   } else {